requirements.txt
-f https://download.pytorch.org/whl/cu113/torch_stable.html
torch==2.2.0
torchvision==0.11.1+cu113
torchaudio==0.10.0+cu113 
numpy>=1.16.5
scipy>=1.3.0
opencv-python
gdown
pillow
gradio
timm==0.4.12
wandb
matplotlib